# api-specs/api-spec.json
{"openapi":"3.0.1","info":{"title":"Employee-API","description":"A REST based API that serves basic information about employees.","version":"0.0.1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/employees/{id}":{"get":{"tags":["employee-controller"],"description":"This method fetches a employee resource from the database based on the given employee id.","operationId":"findEmployeeById","parameters":[{"name":"id","in":"path","description":"The id of the employee to look up.","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Employee requested for the given id doesn't exist in the database.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelEmployee"}}}},"200":{"description":"The employee resource requested has been fetched successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}},"put":{"tags":["employee-controller"],"summary":"This method updates a Employee resource with requisite details and updates it to the database.","operationId":"updateEmployee","parameters":[{"name":"id","in":"path","description":"The id of the employee to update.","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}},"required":true},"responses":{"409":{"description":"Employee data with same update already exists.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelEmployee"}}}},"400":{"description":"Employee data is invalid, e.g. primary emailAddress is missing.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelEmployee"}}}},"200":{"description":"The Employee resource has been updated successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}},"delete":{"tags":["employee-controller"],"summary":"This method deletes a Employee resource from the database for the given employee id.","operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","description":"The id of the employee to delete.","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"The Employee resource has been deleted successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"404":{"description":"Employee data is invalid, e.g. primary emailAddress is missing."}}}},"/employees":{"post":{"tags":["employee-controller"],"summary":"This method creates a Employee resource with requisite details and adds it to the database.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}},"required":true},"responses":{"201":{"description":"The Employee resource has been created successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"400":{"description":"Employee data is invalid, e.g. primary emailAddress is missing.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelEmployee"}}}}}}},"/employees/names/{firstName}":{"get":{"tags":["employee-controller"],"description":"This method fetches a employee resource from the database based on the given employee firstName.","operationId":"findEmployeeByFirstName","parameters":[{"name":"firstName","in":"path","description":"The first name of the employee to look up.","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Employee requested for the given firstName doesn't exist in the database.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelEmployee"}}}},"200":{"description":"The employee resource requested has been fetched successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}}},"/employees/departments/names/{departmentName}":{"get":{"tags":["employee-controller"],"description":"This method fetches a employee resource from the database based on the given employee's department name.","operationId":"findEmployeeByDepartmentName","parameters":[{"name":"departmentName","in":"path","description":"The department name of the employee to look up.","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The employee resource requested has been fetched successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"404":{"description":"Employee requested for the given department name doesn't exist in the database.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelEmployee"}}}}}}}},"components":{"schemas":{"Address":{"required":["city","state"],"type":"object","properties":{"city":{"type":"string"},"street":{"type":"string"},"zipcode":{"type":"integer","format":"int32"},"state":{"type":"string"}}},"Department":{"required":["departmentId","function","name"],"type":"object","properties":{"departmentId":{"type":"integer","format":"int32"},"name":{"type":"string"},"size":{"type":"integer","format":"int32"},"function":{"type":"string"}}},"EmailAddress":{"required":["email"],"type":"object","properties":{"email":{"type":"string"},"primary":{"type":"boolean"}}},"Employee":{"required":["address","department","emailAddresses","firstName","lastName"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"department":{"$ref":"#/components/schemas/Department"},"id":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"address":{"$ref":"#/components/schemas/Address"},"emailAddresses":{"type":"array","items":{"$ref":"#/components/schemas/EmailAddress"}}}},"EntityModelEmployee":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"department":{"$ref":"#/components/schemas/Department"},"id":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"address":{"$ref":"#/components/schemas/Address"},"emailAddresses":{"type":"array","items":{"$ref":"#/components/schemas/EmailAddress"}},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"Link":{"type":"object","properties":{"rel":{"type":"string"},"href":{"type":"string"},"hreflang":{"type":"string"},"media":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"}}},"CollectionModelEntityModelEmployee":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}},"content":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelEmployee"}}}}}}}